requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
jinja2==3.1.4
python-dotenv==1.0.1
httpx==0.27.2
fal-client==0.5.6
pillow==10.4.0
aiofiles==24.1.0
